use graphql.py util
