+++ atd-etl/data_model/env_template
@@ -3,4 +3,4 @@ AWS_ACCESS_KEY_ID=
 AWS_SECRET_ACCESS_KEY=
 BUCKET_NAME=
 EXTRACT_PASSWORD=
-HASURA_GRAPHQL_ENDPOINT=http://localhost:8084/v1/graphql
+HASURA_GRAPHQL_ENDPOINT="http://localhost:8084/v1/graphql"
