+++ atd-etl/data_model/utils/exceptions.py
@@ -1,2 +1,3 @@
 class HasuraAPIError(Exception):
+    """Indicates an error when interacting with the Hasura graphQL API"""
     pass
