+++ atd-etl/data_model/utils/settings.py
@@ -11,3 +11,5 @@ NEW_CR3_FORM_TEST_PIXELS = [
     (625, 1806),
     (4834, 279),
 ]
+
+CSV_UPLOAD_BATCH_SIZE = 1000
